Class simplification according to Scrutinizer
Generic methods of the InvoiceToAccounting class are carried to the AccountingClass parent class according to the recommendations of Scrutinizer.
